requirements.txt
asyncio
web3
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-asyncpg==0.39b0
black==23.1.0
isort==5.12.0
mypy==1.0.0
pylint==2.16.1